// Aeris-Launcher-master/src/app/utils/types/profile.d.ts
export type Profile = {
  accountId: string | null;
  displayName: string | null;
  username: string | null;
  email: string | null;
  password?: string | null;
  discordId: string | null;
  discordAvatar: string | null;
  loginTime: number | null;
  hydrated: boolean;
  validSession: () => boolean;
  isSessionExpired: () => boolean;
  clearSession: () => void;
  clearAllData: () => void;
  isBanned: boolean;
  banReason?: string;
  ticketId?: string;
  lastUsernameChange?: string | null;
  role?: string | null;
  usernameBypass?: boolean;
  checkBanStatus: () => Promise<void>;
  submitAppeal: (ticketId: string, appealText: string) => Promise<void>;
  setProfile: (profile: {
    accountId?: string | null;
    displayName?: string | null;
    username?: string | null;
    email?: string | null;
    password?: string | null;
    discordId?: string | null;
    discordAvatar?: string | null;
    loginTime?: number | null;
    isBanned?: boolean;
    banReason?: string;
    ticketId?: string;
    lastUsernameChange?: string | null;
    role?: string | null;
  }) => void;
  clearProfile: () => void;
  login: (profile: {
    accountId: string;
    displayName: string;
    username?: string;
    email: string | null;
    password: string | null;
    discordId?: string | null;
    discordAvatar?: string | null;
  }) => void;
  logout: () => void;
  setHydrated: () => void;
};